Added console, fixed bugs

+++ Epsilon Engine/include/Texture.h
@@ -125,7 +125,7 @@ public:
         glGenerateMipmap(type);
         glTexParameteri(type, GL_TEXTURE_WRAP_S, wrap);
         glTexParameteri(type, GL_TEXTURE_WRAP_T, wrap);
-        if(filtering  == -1) {
+        if(filtering  == (unsigned int)-1) {
             if(DATA.ANISOTROPY <= 0) {
                 glTexParameteri(type, GL_TEXTURE_MIN_FILTER, GL_LINEAR_MIPMAP_LINEAR);
                 glTexParameteri(type, GL_TEXTURE_MAG_FILTER, GL_LINEAR_MIPMAP_LINEAR);
